+++ bayesflow/diagnostics/metrics/__init__.py
@@ -3,4 +3,4 @@ from .posterior_contraction import posterior_contraction
 from .root_mean_squared_error import root_mean_squared_error
 from .expected_calibration_error import expected_calibration_error
 from .classifier_two_sample_test import classifier_two_sample_test
-from .mmd_hypothesis_test import bootstrap_comparison, summary_space_comparison
+from .model_misspecification import bootstrap_comparison, summary_space_comparison
